lab_6

+++ lab_6/dir-and-files/task10.py
@@ -7,7 +7,12 @@ def check_file_access(path):
     if os.path.exists(path):
         if os.path.isfile(path):
             if os.access(path, os.W_OK) and os.access(path, os.R_OK) and os.access(path, os.X_OK):
-                os.remove(path)
+                return True
+    return False
 
 
-check_file_access(path)
+def remove_file():
+    os.remove(path) if check_file_access(path) else print("File path doesn't exist or access is denied!")
+
+
+remove_file()
